fix after Pierre reviews
