+++ src/app/components/game-session/game-session-card/game-session-card.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input, OnInit } from '@angular/core';
-import { GameSessionModel } from '../../../models/game-session.model';
+import { GameSessionModel } from '../../../models/game-session/game-session.model';
 import { User } from '../../../models/user.model';
 import { GameSessionStatus } from '../../../utils/enum';
