+++ src/app/models/game-session/game-session-create.model.ts
@@ -1,0 +1,6 @@
+export class GameSessionCreateModel {
+    name?: string;
+    game?: string;
+    players?: string[];
+    status?: number;
+}
