+++ src/app/models/game-session/game-session.model.ts
@@ -1,0 +1,12 @@
+import { Game } from '../game.model';
+import { User } from '../user.model';
+import { FormControl } from '@angular/forms';
+
+export class GameSessionModel {
+    name?: string;
+    game?: Game;
+    winner?: string;
+    createdBy?: User;
+    players?: User[];
+    status?: number;
+}
