+++ src/app/pages/list-game-session/list-game-session.component.ts
@@ -4,7 +4,7 @@ import { Collection } from 'ngx-pagination';
 import { ProfilService } from '../../services/profil/profil.service';
 import { MatDialog } from '@angular/material/dialog';
 import { CreateGameSessionDialogComponent } from '../../components/game-session/create-game-session-dialog/create-game-session-dialog.component';
-import { GameSessionModel } from '../../models/game-session.model';
+import { GameSessionModel } from '../../models/game-session/game-session.model';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { AuthService } from '../../services/auth/auth.service';
