+++ src/app/services/game-session/game-session.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { environment } from '../../../environments/environment';
 import { HttpClient } from '@angular/common/http';
-import { GameSessionCreateModel } from '../../models/game-session.model';
+import { GameSessionCreateModel } from '../../models/game-session/game-session-create.model';
 
 const URL = environment.apiUrl + '/game-sessions';
 @Injectable({
